SnapdragonCamera: Fix the issue SoundPoolThread is not released
Two SoundPool objects are created during onCreate. However,
the SoundPool is not released or re-used, properly and app
continuously creating the new one.
Fixing the issue for shutter button SoundPool by adding the
releasing call and changing the call flow so it gets created
and released during onPause and onResume. For count down
SoundPool, we just re-use the existing SoundPool object
instead of creating the new one.

CRs-Fixed: 939366
Change-Id: I8ff1d62ce7b0e217df29ded92c9a07b02bf0ffbb