SnapdragonCamera: Fixed NPE during camera resume
There is a race condition that mMenu is not initialized before
UI thread accessing the object. This can happen if the
initialization of UI thread is completed before camera thread
opens the camera device. Since the camera mMenu instance is
created in the callback of onCameraOpened(), UI thread needs
to check if mMenu is initialized or not before accessing it.

Change-Id: Ic8ffc50b7334714de9a57b22f8229bc0ce868901
CRs-Fixed: 1013999