Bokeh tip: fix visibility
f5a77befb5 introduced a change in which the bokeh tip toolbox is only
visible in certain circumstances. However, that change doesn't take
into account the state of visibility: while in some parts it controls
the visibility depending on its conditions, on other parts it just
shows the tip with all the rest of the bokeh mode.

This patch fixes the tip visibility by saving its state, independently
of wether or not the entire bokeh UI is visible or not.

Change-Id: Id8bd71809d8b17f6c9df15db6a9ffa5f65f41931